+++ lib/tokens.js
