+++ package.json
@@ -24,6 +24,7 @@
     "basic-auth": "^2.0.1",
     "clf-date": "0.2.0",
     "minimist": "^1.2.0",
+    "pino-abstract-transport": "^0.3.0",
     "readable-stream": "^3.6.0",
     "sonic-boom": "^2.2.3",
     "split2": "^3.1.1",
